sandbox/examples/langgraph-agent/pyproject.toml
[project]
name = "langgraph-agent"
version = "0.1.0"
description = "LangGraph agent with agent-sandbox integration using Volcengine API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-sandbox",
    "langgraph>=1.0.0",
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.2",
]

[tool.uv.sources]
agent-sandbox = { path = "../../sdk/python", editable = true }
